June10th_LambdaExample and Predicate
