+++ src/java/RalucaG/MethodsAndEncapsulation/CheckIfHouse.java
@@ -1,0 +1,7 @@
+package MethodsAndEncapsulation;
+
+public class CheckIfHouse implements Cats.CheckType {
+    public boolean types(Cats c) {
+        return c.isHouseCat();
+    }
+}
