+++ src/java/RalucaG/MethodsAndEncapsulation/CheckIfWild.java
@@ -1,0 +1,7 @@
+package MethodsAndEncapsulation;
+
+public class CheckIfWild implements Cats.CheckType {
+    public boolean types(Cats c) {
+        return c.iswildCat();
+    }
+}
